backend/requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pydantic[email]==2.10.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
reportlab==4.2.2

sqlalchemy-utils
